% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_names.R
\name{check_data_names}
\alias{check_data_names}
\title{Check data names}
\usage{
check_data_names(game_id, data)
}
\arguments{
\item{game_id}{The game id, a 64-bit integer.}

\item{data}{The data of \link{data.frame} to check.}
}
\value{
A \link{logical} value indicates whether the data names are correct.
}
\description{
Check if the data names are correct.
}
