% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_prep_fun.R
\docType{data}
\name{config_prep_fun}
\alias{config_prep_fun}
\title{Configurations for the pre-processing functions}
\format{
A list containing input and output variable names for each
pre-processing function. Input variable names are also named lists.
}
\usage{
config_prep_fun
}
\description{
This is used as a default and example configuration for using preproc.iquizoo
package.
}
\keyword{datasets}
