% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.iquizoo-package.R
\docType{package}
\name{data.iquizoo-package}
\alias{data.iquizoo}
\alias{data.iquizoo-package}
\title{data.iquizoo: Useful datasets used in IQUIZOO workflow}
\description{
There will be many dependent dataset for an IQUIZOO workflow to work, such as game configurations of pre-processing functions, abbreviation names, etc. Use a data package to distribute these will be a very great choice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iquizoo/data.iquizoo}
  \item \url{https://iquizoo.github.io/data.iquizoo/}
  \item Report bugs at \url{https://github.com/iquizoo/data.iquizoo/issues}
}

}
\author{
\strong{Maintainer}: Liang Zhang \email{psychelzh@outlook.com} (\href{https://orcid.org/0000-0001-9041-1150}{ORCID})

}
\keyword{internal}
