% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_indices}
\alias{game_indices}
\title{Game indices configuration}
\format{
A data frame with 260 rows of the following variables:
\describe{

\item{game_id}{Game id, a 64-bit integer.}

\item{index_main}{Configure the main index in the raw data.}

\item{index_reverse}{Indicate if the main index score is reversed or not.}
}
}
\usage{
game_indices
}
\description{
The recommended indices for all the games supported by IQUIZOO.COM. Note that
only one index is recommended for each game and some games are not included
with recommended indices.
}
\keyword{datasets}
