% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_info}
\alias{game_info}
\title{Game information}
\format{
A data frame with 350 rows of the following variables:
\describe{

\item{game_id}{Game id, a 64-bit integer.}

\item{game_name}{Game name in Chinese, matching that in the database.}

\item{game_name_ver}{Game version name.}

\item{game_name_en}{English game name.}

\item{game_name_abbr}{English game name abbreviated.}

\item{prep_fun_name}{Name of the preprocessing function.}

\item{input}{Configure input variable names in the raw data.}

\item{extra}{Configure extra special values in the raw data.}
}
}
\usage{
game_info
}
\description{
The basic configurations for all the games supported by IQUIZOO.COM.
}
\keyword{datasets}
