% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_info.R
\name{match_info}
\alias{match_info}
\title{Match specific piece of information of games}
\usage{
match_info(x, from = "game_name_abbr", to = "game_id")
}
\arguments{
\item{x}{The known information of games.}

\item{from}{The name of the column of the known information.}

\item{to}{The name of the column of the information to be matched.}
}
\value{
The matched information.
}
\description{
Match specific piece of information of games
}
