% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_preproc.R
\name{merge_preproc}
\alias{merge_preproc}
\title{Match pre-processing parameters}
\usage{
merge_preproc(
  games,
  filter_only = FALSE,
  rm_tagged = FALSE,
  name_key = "game_id"
)
}
\arguments{
\item{games}{A \link{data.frame} contains the games to match parameters.}

\item{filter_only}{Whether to perform filtering only, i.e., keep only the
matched games.}

\item{rm_tagged}{Whether to remove games with a tagged pre-processing
function. Default is \code{FALSE}.}

\item{name_key}{The name of the column to match the games.}
}
\value{
A \link{data.frame} added pre-processing parameters. Note if there is no
match for a game, the matched parameters are empty vectors (e.g., \code{NULL}).
}
\description{
These complex parameters are stored internally, and users must retrieve them
through this function.
}
