% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_indices.R
\name{screen_indices}
\alias{screen_indices}
\title{Screen out unused indices}
\usage{
screen_indices(data, col_index = "index_name", col_score = "score")
}
\arguments{
\item{data}{A data frame with indices.}

\item{col_index}{The name of the column of the index name.}

\item{col_score}{The name of the column of the index score.}
}
\value{
A data frame with screened indices.
}
\description{
Screen out unused indices
}
