% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_info.R
\name{match_info}
\alias{match_info}
\title{Match specific piece of information of games}
\usage{
match_info(x, to, from = "game_id", fail_on_missing = TRUE)
}
\arguments{
\item{x}{The known information of games.}

\item{to}{The name of the column of the information to be matched.}

\item{from}{The name of the column of the known information.}

\item{fail_on_missing}{Whether to stop if some of the information is not
matched.}
}
\value{
The matched information.
}
\description{
Match specific piece of information of games
}
