% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_preproc.R
\name{match_preproc}
\alias{match_preproc}
\title{Match pre-processing parameters}
\usage{
match_preproc(games)
}
\arguments{
\item{games}{A \link{data.frame} contains the games to match parameters.}
}
\value{
A \link{data.frame} added pre-processing parameters. Note if there is no
match for a game, the matched parameters are empty vectors (e.g., \code{NULL}).
}
\description{
These complex parameters are stored internally, and users must retrieve them
through this function.
}
